moved contactConstraints helpers from util lib to drakeRBM lib
