+++ systems/plants/RigidBodyManipulator.cpp
@@ -2875,111 +2875,6 @@ void RigidBodyManipulator::HandC(MatrixBase<DerivedG> const & q, MatrixBase<Deri
   }
 }
 
-// Builds the part of the contact Jacobian matrix corresponding to bodyInd
-//  by accumulating positive contributions  from points on body A and negative 
-//  contributions from the points on body B.
-//  This is because the normal vectors(n = dphi/dq) are defined to be pointing from B to A
-// INPUTS
-//   bodyInd: the index of the body currently being processed
-//   bodyPoints: (3 x n) matrix where each column is a point on the body
-//   cindA: indexes into the original set of m contact pairs where the body appears as Body A
-//   cindB: indexes into the original set of m contact pairs where the body appears as Body B
-// NOTE 
-//   cindA and cindB are gotten by calling findContactIndexes in drakeContactConstraintsUtil
-//   cols(bodyPoints) = size(cindA) + size(cindB)
-// OUTPUTS:
-//   J: (3m x nq) The partial contact Jacobian matrix
-// NOTE
-//  After one call to the function, the n rows of the Jacobian matrix corresponding to bodyInd will be completed
-//  This function must be called with all bodyInds to finish the total accumulation of the contact Jacobian
-
-void RigidBodyManipulator::accumulateContactJacobian(const int bodyInd, MatrixXd const & bodyPoints, std::vector<int> const & cindA, std::vector<int> const & cindB, MatrixXd & J)
-{
-  const int nq = J.cols();
-  const int numPts = bodyPoints.cols();
-  const size_t numCA = cindA.size();
-  const size_t numCB = cindB.size();
-  const size_t offset = 3*numCA;
-
-  MatrixXd J_tmp(3*numPts, nq);
-  forwardJac(bodyInd - 1, bodyPoints, 0, J_tmp);
-
-  //add contributions from points in xA
-  for (int x = 0 ; x < numCA ; x++) {
-    J.block(3*cindA[x], 0, 3, nq) += J_tmp.block(3*x, 0, 3, nq);
-  }
-
-  //subtract contributions from points in xB
-  for (int x = 0 ; x < numCB ; x++) {
-    J.block(3*cindB[x], 0, 3, nq) -= J_tmp.block(offset + 3*x, 0, 3, nq);
-  }
-}
-
-// Same as above but computes the second order contact Jacobian
-// OUTPUTS:
-//  dJ: (3m x nq^2) Second order contact Jacobian
-void RigidBodyManipulator::accumulateSecondOrderContactJacobian(const int bodyInd, MatrixXd const & bodyPoints, std::vector<int> const & cindA, std::vector<int> const & cindB, MatrixXd & dJ)
-{
-  const int dJCols = dJ.cols(); //nq^2 instead of nq
-  const int numPts = bodyPoints.cols();
-  const size_t numCA = cindA.size();
-  const size_t numCB = cindB.size();
-  const size_t offset = 3*numCA;
-  MatrixXd dJ_tmp(3*numPts, dJCols);
-  forwarddJac(bodyInd - 1, bodyPoints, dJ_tmp); //dJac instead of Jac
-
-  //add contributions from points in xA
-  for (int x = 0 ; x < numCA ; x++) {
-    dJ.block(3*cindA[x], 0, 3, dJCols) += dJ_tmp.block(3*x, 0, 3, dJCols);
-  }
-
-  //subtract contributions from points in xB
-  for (int x = 0 ; x < numCB ; x++) {
-    dJ.block(3*cindB[x], 0, 3, dJCols) -= dJ_tmp.block(offset + 3*x, 0, 3, dJCols);
-  }  
-}
-
-// Computes the full contact Jacobian and, optionally, the second order contact Jacobian.  
-//  This can be used to compute the contact normals in joint coordinates (n = dphi/dq), the surface tangents
-//  in joint coordinates (D), and their respective second derivatives with respect to q (dn, dD)
-// INPUTS
-//   idxA: (m x 1) an integer list of body indexes of body A for m possible contact pairs
-//   idxB: (m x 1) an integeer list of body indexes of body B for m possible contact pairs
-//   xA: (3 x m) each column of the matrix is a contact point in the body A frame for that contact pair
-//   xB: (3 x m) each column of the matrix is a contact point in the body B frame for that contact pair
-//   compute_second_derivatives: boolean flag to indicate that the second order contact Jacobian should also be computed
-// OUTPUTS
-//   J: (3m x nq)
-//  (optional outputs if compute_second_derivatives is true)
-//  dJ: (3m x nq^2) Second order contact Jacobian
-
-void RigidBodyManipulator::computeContactJacobians(Map<VectorXi> const & idxA, Map<VectorXi> const & idxB, Map<Matrix3xd> const & xA, Map<Matrix3xd> const & xB, const bool compute_second_derivatives, MatrixXd & J, MatrixXd & dJ)
-{
-  std::vector<int> bodyInds;
-  const int nq = num_positions;
-  const int numContactPairs = xA.cols();
-
-  J = MatrixXd::Zero(3*numContactPairs, nq);
-  dJ = MatrixXd::Zero(3*numContactPairs, nq*nq);
-  
-  getUniqueBodiesSorted(idxA, idxB, bodyInds);
-  
-  const int numUniqueBodies = bodyInds.size();
-
-  for (int i = 0; i < numUniqueBodies ; i++) {
-    const int bodyInd = bodyInds[i];
-    vector<int> cindA, cindB;
-    MatrixXd bodyPoints;
-    findContactIndexes(idxA, bodyInd, cindA);
-    findContactIndexes(idxB, bodyInd, cindB);
-    getBodyPoints(cindA, cindB, xA, xB, bodyPoints);
-    accumulateContactJacobian(bodyInd, bodyPoints, cindA, cindB, J);
-    if (compute_second_derivatives) {
-      accumulateSecondOrderContactJacobian(bodyInd, bodyPoints, cindA, cindB, dJ);
-    }
-  } 
-}
-
 int RigidBodyManipulator::findLinkId(string linkname, int robot)
 {
   std::transform(linkname.begin(), linkname.end(), linkname.begin(), ::tolower); // convert to lower case
