+++ systems/plants/surfaceTangentsmex.cpp
@@ -1,7 +1,7 @@
 #include "mex.h"
 #include <iostream>
 #include "drakeUtil.h"
-#include "drakeContactConstraintsUtil.h"
+#include "RigidBodyManipulator.h"
 #include <Eigen/Dense>
 #include "math.h"
 
@@ -16,7 +16,7 @@ using namespace std;
  *     surfaceTangents(mex_model_ptr, normals)
  */
 
-inline mxArray* getTangentsArray(Map<Matrix3xd> const & normals)
+inline mxArray* getTangentsArray(RigidBodyManipulator* const model, Map<Matrix3xd> const & normals)
 {
   const int numContactPairs = normals.cols();
   const mwSize cellDims[] = {1, BASIS_VECTOR_HALF_COUNT};
@@ -30,27 +30,29 @@ inline mxArray* getTangentsArray(Map<Matrix3xd> const & normals)
     mxSetCell(tangentCells, k, cell);
   }
 
-  surfaceTangents(normals, tangents);
+  model->surfaceTangents(normals, tangents);
   return tangentCells;
 }
 
 void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) 
 {
-  if (nrhs != 1) {
-    mexErrMsgIdAndTxt("Drake:surfaceTangentsmex:NotEnoughInputs","Usage: [d] = surfaceTangentsmex(normals)");
+  if (nrhs != 2) {
+    mexErrMsgIdAndTxt("Drake:surfaceTangentsmex:NotEnoughInputs","Usage: [d] = surfaceTangentsmex(mex_model_ptr, normals)");
   }
-
-  const unsigned int numNormals = mxGetN(prhs[0]);  //number of normal vectors
-  const unsigned int dimNormals = mxGetM(prhs[0]);  //dimension of each normal vector
   
+  RigidBodyManipulator *model= (RigidBodyManipulator*) getDrakeMexPointer(prhs[0]);
+  const unsigned int numNormals = mxGetN(prhs[1]);  //number of normal vectors
+  const unsigned int dimNormals = mxGetM(prhs[1]);  //dimension of each normal vector
+  
+
   if (dimNormals != 3) {
   	mexErrMsgIdAndTxt("Drake:surfaceTangentsmex:InvalidNormalDimensions","Normals must be 3-Dimensional");	
   }
 
   //Mapping Eigen Matrix to existing memory to avoid copy overhead
-  Map<Matrix3xd> normals(mxGetPr(prhs[0]), 3, numNormals); 
+  Map<Matrix3xd> normals(mxGetPr(prhs[1]), 3, numNormals); 
 
   if (nlhs > 0) {
-    plhs[0] = getTangentsArray(normals);
+    plhs[0] = getTangentsArray(model, normals);
   }
 }
